Revert "octeontx2-pf: Vary completion queue event size"
Until ethtool utility supports changing CQE use devlink
interface only. Hence reverting the patch.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/72996
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>

+++ drivers/net/ethernet/marvell/octeontx2/nic/otx2_common.h
@@ -182,7 +182,8 @@ struct otx2_hw {
 
 #define OTX2_DEFAULT_RBUF_LEN	2048
 	u16			rbuf_len;
-	u32			xqe_size;
+
+	u16			xqe_size;
 
 	/* NPA */
 	u32			stack_pg_ptrs;  /* No of ptrs per stack page */
